reference/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.7
python-dotenv==1.0.0
pyTelegramBotAPI==4.29.1
PlexAPI==4.16.0
matplotlib==3.10.1
pandas==2.0.3
numpy==1.25.0
APScheduler==3.11.0
requests==2.31.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
python-dateutil==2.8.2
pytz==2023.3
tzdata==2023.3
tzlocal==5.3.1
paramiko==3.5.0
lxml==5.4.0
beautifulsoup4==4.12.3
Pillow==10.1.0
openpyxl==3.1.2
schedule==1.2.0
selenium==4.11.2
tabulate==0.9.0
bcrypt==4.2.1
